Can’t make Countable work with Rust’s type system (which might make it so we can’t make the DFAs we need)
If we have two symbol ranges with countable symbols - eg: 1-2 and 3-4 - then they can be considered to 'overlap' in that the range 1-4 will cover exactly the same set of symbols. However, right now we won't detect this.

This is because the ranges 1.0-2.0 and 3.0-4.0 do not overlap because they use uncountable symbols. Rust doesn't seem to have a type system quite smart enough to let us define a version of overlaps that can understand this.