Redeclare the tokenization function so that it has mapping
Not sure how we deal with ‘optional’ values that aren’t in the original. We can treat them as a bunch of isolated streams, strip them, or make it an error.

I think in most cases we actually want to map everything so stripping them or making it an error might work best. However, for the parsing idea I think we actually want the isolated streams option (which is trickier to implement as we have to generate the right ranges for the result and also deal with breaks in the stream)